Deleting Noticias and passing functionalities to News component
